refactor cli by organizaing them into files
